requirements.txt
flask
flask-sqlalchemy
flask-wtf
mysql-connector-python
mysqlclient
flask-Login
email_validator